+++ cmd/zetae2e/local/v2.go
@@ -26,6 +26,8 @@ func startV2Tests(eg *errgroup.Group, conf config.Config, deployerRunner *runner
 		e2etests.TestV2ZEVMToEVMCallName,
 		e2etests.TestV2ZEVMToEVMCallThroughContractName,
 		e2etests.TestV2EVMToZEVMCallName,
+		e2etests.TestV2ETHDepositAndCallNoMessageName,
+		e2etests.TestV2ETHWithdrawAndCallNoMessageName,
 	))
 
 	// Test happy paths for erc20 token workflow
@@ -36,6 +38,7 @@ func startV2Tests(eg *errgroup.Group, conf config.Config, deployerRunner *runner
 		e2etests.TestV2ERC20WithdrawName,
 		e2etests.TestV2ERC20WithdrawAndArbitraryCallName,
 		e2etests.TestV2ERC20WithdrawAndCallName,
+		e2etests.TestV2ERC20DepositAndCallNoMessageName,
 	))
 
 	// Test revert cases for gas token workflow