+++ e2e/e2etests/test_v2_eth_deposit_and_call_no_message.go
@@ -12,26 +12,20 @@ import (
 	crosschaintypes "github.com/zeta-chain/node/x/crosschain/types"
 )
 
-const payloadMessageDepositERC20 = "this is a test ERC20 deposit and call payload"
-
-func TestV2ERC20DepositAndCall(r *runner.E2ERunner, args []string) {
+func TestV2ETHDepositAndCallNoMessage(r *runner.E2ERunner, args []string) {
 	require.Len(r, args, 1)
 
 	amount, ok := big.NewInt(0).SetString(args[0], 10)
-	require.True(r, ok, "Invalid amount specified for TestV2ERC20DepositAndCall")
+	require.True(r, ok, "Invalid amount specified for TestV2ETHDepositAndCallNoMessage")
 
-	r.ApproveERC20OnEVM(r.GatewayEVMAddr)
-
-	r.AssertTestDAppZEVMCalled(false, payloadMessageDepositERC20, amount)
-
-	oldBalance, err := r.ERC20ZRC20.BalanceOf(&bind.CallOpts{}, r.TestDAppV2ZEVMAddr)
+	oldBalance, err := r.ETHZRC20.BalanceOf(&bind.CallOpts{}, r.TestDAppV2ZEVMAddr)
 	require.NoError(r, err)
 
-	// perform the deposit
-	tx := r.V2ERC20DepositAndCall(
+	// perform the deposit and call to the TestDAppV2ZEVMAddr
+	tx := r.V2ETHDepositAndCall(
 		r.TestDAppV2ZEVMAddr,
 		amount,
-		[]byte(payloadMessageDepositERC20),
+		[]byte{},
 		gatewayevm.RevertOptions{OnRevertGasLimit: big.NewInt(0)},
 	)
 
@@ -41,10 +35,12 @@ func TestV2ERC20DepositAndCall(r *runner.E2ERunner, args []string) {
 	require.Equal(r, crosschaintypes.CctxStatus_OutboundMined, cctx.CctxStatus.Status)
 
 	// check the payload was received on the contract
-	r.AssertTestDAppZEVMCalled(true, payloadMessageDepositERC20, amount)
+	messageIndex, err := r.TestDAppV2ZEVM.GetNoMessageIndex(&bind.CallOpts{}, r.EVMAddress())
+	require.NoError(r, err)
+	r.AssertTestDAppZEVMCalled(true, messageIndex, amount)
 
 	// check the balance was updated
-	newBalance, err := r.ERC20ZRC20.BalanceOf(&bind.CallOpts{}, r.TestDAppV2ZEVMAddr)
+	newBalance, err := r.ETHZRC20.BalanceOf(&bind.CallOpts{}, r.TestDAppV2ZEVMAddr)
 	require.NoError(r, err)
 	require.Equal(r, new(big.Int).Add(oldBalance, amount), newBalance)
 }